Scripts I have made
Each of these files has some functionality that will be used in the final code. Just need to copy/paste and format them correctly, then add a few other bits of code. A path to python IDE is needed, or finding the IDE files, and a function to write to them, creating a new theme possibly making that theme the theme that is being used. So color number has to be written to the file. Perhaps a drop down menu with the colors in the file, backrounf, text ect. Then once you picked a color (showing on button or where ever) you choose in a drop down or check box, radio button, anything where that color should be written. Should the colors be written to a seperate file, and then that written into IDE files or write directly to IDE file?
user may have to enter their path as python can be installed in many places.

First lets just write the colors to a file. Then worry about the order, and getting them to Python IDE.